requirements.txt
aiogram==3.7.0
sqlalchemy[asyncio]==2.0.29
aiosqlite==0.20.0
asyncpg==0.29.0
python-dotenv==1.0.1
openpyxl==3.1.2